login,signup style refactoring / fix logo position
