+++ client/src/App.jsx
@@ -1,11 +1,9 @@
 import GlobalStyle from "@/styles/global";
-import QuestionDetail from "@/pages/QuestionDetail";
 
 function App() {
   return (
     <>
       <GlobalStyle />
-      <QuestionDetail />
     </>
   );
 }
