+++ client/src/config/config.jsx
@@ -1,4 +1,4 @@
-export const BrandLogoImg = "../../src/assets/image/pngwing.png";
-export const logoImg = "../../src/assets/image/logo.png";
-export const userProfileImg =
+export const BRANDLOGOIMG = "../../src/assets/image/pngwing.png";
+export const LOGOIMG = "../../src/assets/image/logo.png";
+export const USERPROFILEIMG =
   "https://cdn.pixabay.com/photo/2023/05/05/21/00/cute-7973191_1280.jpg";
