+++ client/src/styles/Login/LoginFormStyle.jsx
@@ -1,8 +1,6 @@
-import { useState } from "react";
-import styled from "styled-components";
-import GlobalStyle from "../styles/global";
+import { styled } from "styled-components";
 
-const LoginContainer = styled.div`
+export const LoginContainer = styled.div`
   width: 100%;
   max-width: 288px;
   height: 234px;
@@ -49,28 +47,3 @@ const LoginContainer = styled.div`
     margin-bottom: 20px;
   }
 `;
-
-const LoginForm = () => {
-  // const [loginInfo, setLoginInfo] = useState({ email: "", password: "" });
-
-  return (
-    <>
-      <GlobalStyle />
-      <LoginContainer className="login-container">
-        <div className="login-email">
-          <label htmlFor="email">Email</label>
-          <input id="email" type="email" />
-        </div>
-        <div className="login-password">
-          <label htmlFor="email">Password</label>
-          <input id="password" type="password" />
-        </div>
-        <div>
-          <button>Log in</button>
-        </div>
-      </LoginContainer>
-    </>
-  );
-};
-
-export default LoginForm;
