+++ client/src/styles/Login/LoginPageStyle.jsx
@@ -1,0 +1,25 @@
+import { styled } from "styled-components";
+
+export const PageContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: 100%;
+  background-color: var(--black-050);
+
+  p {
+    margin: 2px;
+    padding: 2px;
+    color: var(--black);
+    font-size: 12px;
+  }
+`;
+
+export const Icon = styled.img`
+  width: 40px;
+  height: 40px;
+  display: flex;
+  margin: 0 auto 20px;
+`;
